gallery

+++ memories/src/main/java/com/traveljar/memories/audio/DownloadAudioAsyncTask.java
@@ -1,5 +1,6 @@
 package com.traveljar.memories.audio;
 
+import android.content.Context;
 import android.os.AsyncTask;
 import android.util.Log;
 
@@ -7,6 +8,7 @@ import com.traveljar.memories.customevents.AudioDownloadEvent;
 import com.traveljar.memories.models.Audio;
 import com.traveljar.memories.utility.Constants;
 
+import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
@@ -21,6 +23,7 @@ public class DownloadAudioAsyncTask extends AsyncTask<String, Integer, String> {
     private static final String TAG = "DOWNLOAD_ASYNC_TASK";
     private Audio mAudio;
     private final int CALLING_ACTIVITY_CODE;
+    private Context context;
 
     public DownloadAudioAsyncTask(int activityCode, Audio audio) {
         CALLING_ACTIVITY_CODE = activityCode;
@@ -33,47 +36,52 @@ public class DownloadAudioAsyncTask extends AsyncTask<String, Integer, String> {
         InputStream input = null;
         OutputStream output = null;
         HttpURLConnection connection = null;
-        String fileLocation = Constants.TRAVELJAR_FOLDER_AUDIO + System.currentTimeMillis()
-                + ".mp3";
-        URL downloadUrl;
-        try {
-            downloadUrl = new URL(mAudio.getDataServerURL());
-            Log.d(TAG, "started downloading");
-            connection = (HttpURLConnection) downloadUrl.openConnection();
-            connection.connect();
+        String fileLocation = Constants.TRAVELJAR_FOLDER_AUDIO + "/aud_" + mAudio.getCreatedBy() + "_" + mAudio.getjId() +
+                "_" + mAudio.getCreatedAt() + ".3gp";
 
-            // expect HTTP 200 OK, so we don't mistakenly save error report instead of the file
-            if (connection.getResponseCode() != HttpURLConnection.HTTP_OK) {
-                return "Server returned HTTP " + connection.getResponseCode() + " "
-                        + connection.getResponseMessage();
+        if(!(new File(fileLocation)).exists()){
+            try {
+                URL downloadUrl = new URL(mAudio.getDataServerURL());
+                Log.d(TAG, "started downloading");
+                connection = (HttpURLConnection) downloadUrl.openConnection();
+                connection.connect();
+
+                // expect HTTP 200 OK, so we don't mistakenly save error report instead of the file
+                if (connection.getResponseCode() != HttpURLConnection.HTTP_OK) {
+                    return "Server returned HTTP " + connection.getResponseCode() + " "
+                            + connection.getResponseMessage();
+                }
+
+                // download the file
+                input = connection.getInputStream();
+                output = new FileOutputStream(fileLocation);
+
+                byte data[] = new byte[4096];
+                int count;
+                while ((count = input.read(data)) != -1) {
+                    output.write(data, 0, count);
+                }
+                Log.d(TAG, "download finished");
+                mAudio.setDataLocalURL(fileLocation);
+                return fileLocation;
+            } catch (Exception e) {
+                Log.d(TAG, "exception in downloading audio " + e);
+                e.printStackTrace();
+            } finally {
+                try {
+                    if (output != null)
+                        output.close();
+                    if (input != null)
+                        input.close();
+                } catch (IOException ignored) {
+                }
+
+                if (connection != null)
+                    connection.disconnect();
             }
-
-            // download the file
-            input = connection.getInputStream();
-            output = new FileOutputStream(fileLocation);
-
-            byte data[] = new byte[4096];
-            int count;
-            while ((count = input.read(data)) != -1) {
-                output.write(data, 0, count);
-            }
-            Log.d(TAG, "download finished");
+        }else {
             mAudio.setDataLocalURL(fileLocation);
             return fileLocation;
-        } catch (Exception e) {
-            Log.d(TAG, "exception in downloading audio " + e);
-            e.printStackTrace();
-        } finally {
-            try {
-                if (output != null)
-                    output.close();
-                if (input != null)
-                    input.close();
-            } catch (IOException ignored) {
-            }
-
-            if (connection != null)
-                connection.disconnect();
         }
         return null;
     }
