+++ memories/src/main/java/com/traveljar/memories/video/DownloadVideoAsyncTask.java
@@ -1,21 +1,18 @@
 package com.traveljar.memories.video;
 
 import android.os.AsyncTask;
+import android.os.Environment;
 import android.util.Log;
 
 import com.traveljar.memories.models.Video;
-import com.traveljar.memories.utility.Constants;
 
+import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.net.HttpURLConnection;
 import java.net.URL;
-
-/**
- * Created by ankit on 2/6/15.
- */
 
 public class DownloadVideoAsyncTask extends AsyncTask<String, Integer, String> {
 
@@ -34,48 +31,53 @@ public class DownloadVideoAsyncTask extends AsyncTask<String, Integer, String> {
         InputStream input = null;
         OutputStream output = null;
         HttpURLConnection connection = null;
-        String fileLocation = Constants.TRAVELJAR_FOLDER_VIDEO + System.currentTimeMillis() + ".mp4";
-        URL downloadUrl;
-        try {
-            Log.d(TAG, "video server url is " + mVideo.getDataServerURL());
-            downloadUrl = new URL(mVideo.getDataServerURL());
-            Log.d(TAG, "started downloading video");
-            connection = (HttpURLConnection) downloadUrl.openConnection();
-            connection.connect();
-
-            // expect HTTP 200 OK, so we don't mistakenly save error report instead of the file
-            if (connection.getResponseCode() != HttpURLConnection.HTTP_OK) {
-                return "Server returned HTTP " + connection.getResponseCode() + " "
-                        + connection.getResponseMessage();
-            }
-
-            // download the file
-            input = connection.getInputStream();
-            output = new FileOutputStream(fileLocation);
-
-            byte data[] = new byte[4096];
-            int count;
-            while ((count = input.read(data)) != -1) {
-                output.write(data, 0, count);
-            }
-            mVideo.setDataLocalURL(fileLocation);
-            Log.d(TAG, "finished Downloading video");
-            return fileLocation;
-        } catch (Exception e) {
-            Log.d(TAG, "Error in downloading video");
-            e.printStackTrace();
-            return null;
-        } finally {
+        String fileLocation = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_MOVIES).getAbsolutePath() +
+                "/vid_" + mVideo.getCreatedBy() + "_" + mVideo.getjId() + "_" + mVideo.getCreatedAt() + ".mp4";
+        if(!(new File(fileLocation)).exists()) {
             try {
-                if (output != null)
-                    output.close();
-                if (input != null)
-                    input.close();
-            } catch (IOException ignored) {
-            }
+                Log.d(TAG, "video server url is " + mVideo.getDataServerURL());
+                URL downloadUrl = new URL(mVideo.getDataServerURL());
+                Log.d(TAG, "started downloading video");
+                connection = (HttpURLConnection) downloadUrl.openConnection();
+                connection.connect();
 
-            if (connection != null)
-                connection.disconnect();
+                // expect HTTP 200 OK, so we don't mistakenly save error report instead of the file
+                if (connection.getResponseCode() != HttpURLConnection.HTTP_OK) {
+                    return "Server returned HTTP " + connection.getResponseCode() + " "
+                            + connection.getResponseMessage();
+                }
+
+                // download the file
+                input = connection.getInputStream();
+                output = new FileOutputStream(fileLocation);
+
+                byte data[] = new byte[4096];
+                int count;
+                while ((count = input.read(data)) != -1) {
+                    output.write(data, 0, count);
+                }
+                mVideo.setDataLocalURL(fileLocation);
+                Log.d(TAG, "finished Downloading video");
+                return fileLocation;
+            } catch (Exception e) {
+                Log.d(TAG, "Error in downloading video");
+                e.printStackTrace();
+                return null;
+            } finally {
+                try {
+                    if (output != null)
+                        output.close();
+                    if (input != null)
+                        input.close();
+                } catch (IOException ignored) {
+                }
+
+                if (connection != null)
+                    connection.disconnect();
+            }
+        }else {
+            mVideo.setDataLocalURL(fileLocation);
+            return fileLocation;
         }
     }
